fix: rename container migrators
